backend/requirements.txt
# --- Web Framework & Server ---
fastapi
uvicorn[standard]

# --- Environment Variables (For your API Key later) ---
python-dotenv
pydantic